exception-free parsing
