+++ modules/javafx.graphics/src/main/java/javafx/scene/paint/Color.java
@@ -25,10 +25,8 @@
 
 package javafx.scene.paint;
 
-import java.util.Map;
 import java.util.Objects;
-import java.util.TreeMap;
-import com.sun.javafx.css.parser.CssNumberParser;
+import com.sun.javafx.css.parser.CssColorParser;
 import com.sun.javafx.util.InterpolationUtils;
 import com.sun.javafx.util.Utils;
 import com.sun.javafx.tk.Toolkit;
@@ -392,187 +390,15 @@ public final class Color extends Paint {
      */
     public static Color web(String colorString, double opacity) {
         if (colorString == null) {
-            throw new NullPointerException(
-                    "The color components or name must be specified");
-        }
-        if (colorString.isEmpty()) {
-            throw new IllegalArgumentException("Invalid color specification");
+            throw new NullPointerException("Color components or name must be specified");
         }
 
-        int offset = 0;
-
-        if (colorString.charAt(0) == '#') {
-            offset = 1;
-        } else if (colorString.regionMatches(true, 0, "0x", 0, 2)) {
-            offset = 2;
-        } else if (colorString.regionMatches(true, 0, "rgb", 0, 3)) {
-            if (colorString.charAt(3) == '(') {
-                return parseRGBColor(colorString, 4, false, opacity);
-            } else if (colorString.regionMatches(true, 3, "a(", 0, 2)) {
-                return parseRGBColor(colorString, 5, true, opacity);
-            }
-        } else if (colorString.regionMatches(true, 0, "hsl", 0, 3)) {
-            if (colorString.charAt(3) == '(') {
-                return parseHSLColor(colorString, 4, false, opacity);
-            } else if (colorString.regionMatches(true, 3, "a(", 0, 2)) {
-                return parseHSLColor(colorString, 5, true, opacity);
-            }
-        } else {
-            Color col = NamedColors.get(colorString);
-            if (col != null) {
-                if (opacity == 1.0) {
-                    return col;
-                } else {
-                    return Color.color(col.red, col.green, col.blue, opacity);
-                }
-            }
+        Color result = CssColorParser.tryParseColor(colorString, opacity);
+        if (result == null) {
+            throw new IllegalArgumentException("Invalid color: " + colorString);
         }
 
-        int len = colorString.length() - offset;
-
-        try {
-            return switch (len) {
-                case 3 -> {
-                    int r = Integer.parseInt(colorString, offset, offset + 1, 16);
-                    int g = Integer.parseInt(colorString, offset + 1, offset + 2, 16);
-                    int b = Integer.parseInt(colorString, offset + 2, offset + 3, 16);
-                    yield Color.color(r / 15.0, g / 15.0, b / 15.0, opacity);
-                }
-
-                case 4 -> {
-                    int r = Integer.parseInt(colorString, offset, offset + 1, 16);
-                    int g = Integer.parseInt(colorString, offset + 1, offset + 2, 16);
-                    int b = Integer.parseInt(colorString, offset + 2, offset + 3, 16);
-                    int a = Integer.parseInt(colorString, offset + 3, offset + 4, 16);
-                    yield Color.color(r / 15.0, g / 15.0, b / 15.0, opacity * a / 15.0);
-                }
-
-                case 6 -> {
-                    int r = Integer.parseInt(colorString, offset, offset + 2, 16);
-                    int g = Integer.parseInt(colorString, offset + 2, offset + 4, 16);
-                    int b = Integer.parseInt(colorString, offset + 4, offset + 6, 16);
-                    yield Color.rgb(r, g, b, opacity);
-                }
-
-                case 8 -> {
-                    int r = Integer.parseInt(colorString, offset, offset + 2, 16);
-                    int g = Integer.parseInt(colorString, offset + 2, offset + 4, 16);
-                    int b = Integer.parseInt(colorString, offset + 4, offset + 6, 16);
-                    int a = Integer.parseInt(colorString, offset + 6, offset + 8, 16);
-                    yield Color.rgb(r, g, b, opacity * a / 255.0);
-                }
-
-                default -> throw new IllegalArgumentException("Invalid color specification");
-            };
-        } catch (NumberFormatException _) {
-            throw new IllegalArgumentException("Invalid color specification");
-        }
-    }
-
-    private static Color parseRGBColor(String color, int roff,
-                                       boolean hasAlpha, double a)
-    {
-        try {
-            int rend = color.indexOf(',', roff);
-            int gend = rend < 0 ? -1 : color.indexOf(',', rend+1);
-            int bend = gend < 0 ? -1 : color.indexOf(hasAlpha ? ',' : ')', gend+1);
-            int aend = hasAlpha ? (bend < 0 ? -1 : color.indexOf(')', bend+1)) : bend;
-            if (aend >= 0) {
-                double r = parseComponent(color, roff, rend, PARSE_COMPONENT);
-                double g = parseComponent(color, rend+1, gend, PARSE_COMPONENT);
-                double b = parseComponent(color, gend+1, bend, PARSE_COMPONENT);
-                if (hasAlpha) {
-                    a *= parseComponent(color, bend+1, aend, PARSE_ALPHA);
-                }
-                return new Color(r, g, b, a);
-            }
-        } catch (NumberFormatException nfe) {}
-
-        throw new IllegalArgumentException("Invalid color specification");
-    }
-
-    private static Color parseHSLColor(String color, int hoff,
-                                       boolean hasAlpha, double a)
-    {
-        try {
-            int hend = color.indexOf(',', hoff);
-            int send = hend < 0 ? -1 : color.indexOf(',', hend+1);
-            int lend = send < 0 ? -1 : color.indexOf(hasAlpha ? ',' : ')', send+1);
-            int aend = hasAlpha ? (lend < 0 ? -1 : color.indexOf(')', lend+1)) : lend;
-            if (aend >= 0) {
-                double h = parseComponent(color, hoff, hend, PARSE_ANGLE);
-                double s = parseComponent(color, hend+1, send, PARSE_PERCENT);
-                double l = parseComponent(color, send+1, lend, PARSE_PERCENT);
-                if (hasAlpha) {
-                    a *= parseComponent(color, lend+1, aend, PARSE_ALPHA);
-                }
-                return Color.hsb(h, s, l, a);
-            }
-        } catch (NumberFormatException nfe) {}
-
-        throw new IllegalArgumentException("Invalid color specification");
-    }
-
-    private static final int PARSE_COMPONENT = 0; // percent, or clamped to [0,255] => [0,1]
-    private static final int PARSE_PERCENT = 1; // clamped to [0,100]% => [0,1]
-    private static final int PARSE_ANGLE = 2; // clamped to [0,360]
-    private static final int PARSE_ALPHA = 3; // clamped to [0.0,1.0]
-
-    private static double parseComponent(String color, int off, int end, int type) {
-        int start = off;
-        int limit = end;
-
-        while (start < limit && Character.isWhitespace(color.charAt(start))) {
-            start++;
-        }
-
-        while (limit > start && Character.isWhitespace(color.charAt(limit - 1))) {
-            limit--;
-        }
-
-        if (start >= limit) {
-            throw new IllegalArgumentException("Invalid color specification");
-        }
-
-        if (color.charAt(limit - 1) == '%') {
-            if (type > PARSE_PERCENT) {
-                throw new IllegalArgumentException("Invalid color specification");
-            }
-
-            type = PARSE_PERCENT;
-            limit--;
-
-            while (limit > start && Character.isWhitespace(color.charAt(limit - 1))) {
-                limit--;
-            }
-
-            if (start >= limit) {
-                throw new IllegalArgumentException("Invalid color specification");
-            }
-        } else if (type == PARSE_PERCENT) {
-            throw new IllegalArgumentException("Invalid color specification");
-        }
-
-        double c = (type == PARSE_COMPONENT)
-            ? Integer.parseInt(color, start, limit, 10)
-            : CssNumberParser.parseDouble(color, start, limit);
-
-        switch (type) {
-            case PARSE_ALPHA:
-                return (c < 0.0) ? 0.0 : ((c > 1.0) ? 1.0 : c);
-            case PARSE_PERCENT:
-                return (c <= 0.0) ? 0.0 : ((c >= 100.0) ? 1.0 : (c / 100.0));
-            case PARSE_COMPONENT:
-                return (c <= 0.0) ? 0.0 : ((c >= 255.0) ? 1.0 : (c / 255.0));
-            case PARSE_ANGLE:
-                return ((c < 0.0)
-                        ? ((c % 360.0) + 360.0)
-                        : ((c > 360.0)
-                            ? (c % 360.0)
-                            : c));
-        }
-
-        throw new IllegalArgumentException("Invalid color specification");
+        return result;
     }
 
     /**
@@ -1706,170 +1532,6 @@ public final class Color extends Paint {
      * <div style="border:1px solid black;width:40px;height:20px;background-color:#9ACD32;float:right;margin: 0 10px 0 0"></div>
      */
     public static final Color YELLOWGREEN = new Color(0.6039216f, 0.8039216f, 0.19607843f);
-
-    /*
-     * Named colors moved to nested class to initialize them only when they
-     * are needed.
-     */
-    private static final class NamedColors {
-
-        private static Color get(String name) {
-            return NAMED_COLORS.get(name);
-        }
-
-        private static final Map<String, Color> NAMED_COLORS = new TreeMap<>(String.CASE_INSENSITIVE_ORDER);
-
-        static  {
-            NAMED_COLORS.put("aliceblue",            ALICEBLUE);
-            NAMED_COLORS.put("antiquewhite",         ANTIQUEWHITE);
-            NAMED_COLORS.put("aqua",                 AQUA);
-            NAMED_COLORS.put("aquamarine",           AQUAMARINE);
-            NAMED_COLORS.put("azure",                AZURE);
-            NAMED_COLORS.put("beige",                BEIGE);
-            NAMED_COLORS.put("bisque",               BISQUE);
-            NAMED_COLORS.put("black",                BLACK);
-            NAMED_COLORS.put("blanchedalmond",       BLANCHEDALMOND);
-            NAMED_COLORS.put("blue",                 BLUE);
-            NAMED_COLORS.put("blueviolet",           BLUEVIOLET);
-            NAMED_COLORS.put("brown",                BROWN);
-            NAMED_COLORS.put("burlywood",            BURLYWOOD);
-            NAMED_COLORS.put("cadetblue",            CADETBLUE);
-            NAMED_COLORS.put("chartreuse",           CHARTREUSE);
-            NAMED_COLORS.put("chocolate",            CHOCOLATE);
-            NAMED_COLORS.put("coral",                CORAL);
-            NAMED_COLORS.put("cornflowerblue",       CORNFLOWERBLUE);
-            NAMED_COLORS.put("cornsilk",             CORNSILK);
-            NAMED_COLORS.put("crimson",              CRIMSON);
-            NAMED_COLORS.put("cyan",                 CYAN);
-            NAMED_COLORS.put("darkblue",             DARKBLUE);
-            NAMED_COLORS.put("darkcyan",             DARKCYAN);
-            NAMED_COLORS.put("darkgoldenrod",        DARKGOLDENROD);
-            NAMED_COLORS.put("darkgray",             DARKGRAY);
-            NAMED_COLORS.put("darkgreen",            DARKGREEN);
-            NAMED_COLORS.put("darkgrey",             DARKGREY);
-            NAMED_COLORS.put("darkkhaki",            DARKKHAKI);
-            NAMED_COLORS.put("darkmagenta",          DARKMAGENTA);
-            NAMED_COLORS.put("darkolivegreen",       DARKOLIVEGREEN);
-            NAMED_COLORS.put("darkorange",           DARKORANGE);
-            NAMED_COLORS.put("darkorchid",           DARKORCHID);
-            NAMED_COLORS.put("darkred",              DARKRED);
-            NAMED_COLORS.put("darksalmon",           DARKSALMON);
-            NAMED_COLORS.put("darkseagreen",         DARKSEAGREEN);
-            NAMED_COLORS.put("darkslateblue",        DARKSLATEBLUE);
-            NAMED_COLORS.put("darkslategray",        DARKSLATEGRAY);
-            NAMED_COLORS.put("darkslategrey",        DARKSLATEGREY);
-            NAMED_COLORS.put("darkturquoise",        DARKTURQUOISE);
-            NAMED_COLORS.put("darkviolet",           DARKVIOLET);
-            NAMED_COLORS.put("deeppink",             DEEPPINK);
-            NAMED_COLORS.put("deepskyblue",          DEEPSKYBLUE);
-            NAMED_COLORS.put("dimgray",              DIMGRAY);
-            NAMED_COLORS.put("dimgrey",              DIMGREY);
-            NAMED_COLORS.put("dodgerblue",           DODGERBLUE);
-            NAMED_COLORS.put("firebrick",            FIREBRICK);
-            NAMED_COLORS.put("floralwhite",          FLORALWHITE);
-            NAMED_COLORS.put("forestgreen",          FORESTGREEN);
-            NAMED_COLORS.put("fuchsia",              FUCHSIA);
-            NAMED_COLORS.put("gainsboro",            GAINSBORO);
-            NAMED_COLORS.put("ghostwhite",           GHOSTWHITE);
-            NAMED_COLORS.put("gold",                 GOLD);
-            NAMED_COLORS.put("goldenrod",            GOLDENROD);
-            NAMED_COLORS.put("gray",                 GRAY);
-            NAMED_COLORS.put("green",                GREEN);
-            NAMED_COLORS.put("greenyellow",          GREENYELLOW);
-            NAMED_COLORS.put("grey",                 GREY);
-            NAMED_COLORS.put("honeydew",             HONEYDEW);
-            NAMED_COLORS.put("hotpink",              HOTPINK);
-            NAMED_COLORS.put("indianred",            INDIANRED);
-            NAMED_COLORS.put("indigo",               INDIGO);
-            NAMED_COLORS.put("ivory",                IVORY);
-            NAMED_COLORS.put("khaki",                KHAKI);
-            NAMED_COLORS.put("lavender",             LAVENDER);
-            NAMED_COLORS.put("lavenderblush",        LAVENDERBLUSH);
-            NAMED_COLORS.put("lawngreen",            LAWNGREEN);
-            NAMED_COLORS.put("lemonchiffon",         LEMONCHIFFON);
-            NAMED_COLORS.put("lightblue",            LIGHTBLUE);
-            NAMED_COLORS.put("lightcoral",           LIGHTCORAL);
-            NAMED_COLORS.put("lightcyan",            LIGHTCYAN);
-            NAMED_COLORS.put("lightgoldenrodyellow", LIGHTGOLDENRODYELLOW);
-            NAMED_COLORS.put("lightgray",            LIGHTGRAY);
-            NAMED_COLORS.put("lightgreen",           LIGHTGREEN);
-            NAMED_COLORS.put("lightgrey",            LIGHTGREY);
-            NAMED_COLORS.put("lightpink",            LIGHTPINK);
-            NAMED_COLORS.put("lightsalmon",          LIGHTSALMON);
-            NAMED_COLORS.put("lightseagreen",        LIGHTSEAGREEN);
-            NAMED_COLORS.put("lightskyblue",         LIGHTSKYBLUE);
-            NAMED_COLORS.put("lightslategray",       LIGHTSLATEGRAY);
-            NAMED_COLORS.put("lightslategrey",       LIGHTSLATEGREY);
-            NAMED_COLORS.put("lightsteelblue",       LIGHTSTEELBLUE);
-            NAMED_COLORS.put("lightyellow",          LIGHTYELLOW);
-            NAMED_COLORS.put("lime",                 LIME);
-            NAMED_COLORS.put("limegreen",            LIMEGREEN);
-            NAMED_COLORS.put("linen",                LINEN);
-            NAMED_COLORS.put("magenta",              MAGENTA);
-            NAMED_COLORS.put("maroon",               MAROON);
-            NAMED_COLORS.put("mediumaquamarine",     MEDIUMAQUAMARINE);
-            NAMED_COLORS.put("mediumblue",           MEDIUMBLUE);
-            NAMED_COLORS.put("mediumorchid",         MEDIUMORCHID);
-            NAMED_COLORS.put("mediumpurple",         MEDIUMPURPLE);
-            NAMED_COLORS.put("mediumseagreen",       MEDIUMSEAGREEN);
-            NAMED_COLORS.put("mediumslateblue",      MEDIUMSLATEBLUE);
-            NAMED_COLORS.put("mediumspringgreen",    MEDIUMSPRINGGREEN);
-            NAMED_COLORS.put("mediumturquoise",      MEDIUMTURQUOISE);
-            NAMED_COLORS.put("mediumvioletred",      MEDIUMVIOLETRED);
-            NAMED_COLORS.put("midnightblue",         MIDNIGHTBLUE);
-            NAMED_COLORS.put("mintcream",            MINTCREAM);
-            NAMED_COLORS.put("mistyrose",            MISTYROSE);
-            NAMED_COLORS.put("moccasin",             MOCCASIN);
-            NAMED_COLORS.put("navajowhite",          NAVAJOWHITE);
-            NAMED_COLORS.put("navy",                 NAVY);
-            NAMED_COLORS.put("oldlace",              OLDLACE);
-            NAMED_COLORS.put("olive",                OLIVE);
-            NAMED_COLORS.put("olivedrab",            OLIVEDRAB);
-            NAMED_COLORS.put("orange",               ORANGE);
-            NAMED_COLORS.put("orangered",            ORANGERED);
-            NAMED_COLORS.put("orchid",               ORCHID);
-            NAMED_COLORS.put("palegoldenrod",        PALEGOLDENROD);
-            NAMED_COLORS.put("palegreen",            PALEGREEN);
-            NAMED_COLORS.put("paleturquoise",        PALETURQUOISE);
-            NAMED_COLORS.put("palevioletred",        PALEVIOLETRED);
-            NAMED_COLORS.put("papayawhip",           PAPAYAWHIP);
-            NAMED_COLORS.put("peachpuff",            PEACHPUFF);
-            NAMED_COLORS.put("peru",                 PERU);
-            NAMED_COLORS.put("pink",                 PINK);
-            NAMED_COLORS.put("plum",                 PLUM);
-            NAMED_COLORS.put("powderblue",           POWDERBLUE);
-            NAMED_COLORS.put("purple",               PURPLE);
-            NAMED_COLORS.put("red",                  RED);
-            NAMED_COLORS.put("rosybrown",            ROSYBROWN);
-            NAMED_COLORS.put("royalblue",            ROYALBLUE);
-            NAMED_COLORS.put("saddlebrown",          SADDLEBROWN);
-            NAMED_COLORS.put("salmon",               SALMON);
-            NAMED_COLORS.put("sandybrown",           SANDYBROWN);
-            NAMED_COLORS.put("seagreen",             SEAGREEN);
-            NAMED_COLORS.put("seashell",             SEASHELL);
-            NAMED_COLORS.put("sienna",               SIENNA);
-            NAMED_COLORS.put("silver",               SILVER);
-            NAMED_COLORS.put("skyblue",              SKYBLUE);
-            NAMED_COLORS.put("slateblue",            SLATEBLUE);
-            NAMED_COLORS.put("slategray",            SLATEGRAY);
-            NAMED_COLORS.put("slategrey",            SLATEGREY);
-            NAMED_COLORS.put("snow",                 SNOW);
-            NAMED_COLORS.put("springgreen",          SPRINGGREEN);
-            NAMED_COLORS.put("steelblue",            STEELBLUE);
-            NAMED_COLORS.put("tan",                  TAN);
-            NAMED_COLORS.put("teal",                 TEAL);
-            NAMED_COLORS.put("thistle",              THISTLE);
-            NAMED_COLORS.put("tomato",               TOMATO);
-            NAMED_COLORS.put("transparent",          TRANSPARENT);
-            NAMED_COLORS.put("turquoise",            TURQUOISE);
-            NAMED_COLORS.put("violet",               VIOLET);
-            NAMED_COLORS.put("wheat",                WHEAT);
-            NAMED_COLORS.put("white",                WHITE);
-            NAMED_COLORS.put("whitesmoke",           WHITESMOKE);
-            NAMED_COLORS.put("yellow",               YELLOW);
-            NAMED_COLORS.put("yellowgreen",          YELLOWGREEN);
-        }
-    }
 
     /**
      * The red component of the {@code Color}, in the range {@code 0.0-1.0}.
